Remove old oc-based version
